Smaller patch (#3440)
# Motivation
The patch files used when generating rust code from did files are
unnecessarily large. All the code in the headers generated by `didc`
must be commented out, as it conflicts with our own headers. This is
easy to do automatically, but some of lines are still commented out by
verbose and fragile patch files.

# Changes
- Comment out the remaining `didc` header lines automatically.
- Regenerate the patch files.  The new patch files are smaller.

# Tests
There is no change in functionality.  CI should suffice.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>