GIX-1930: Use NF and direct commitment fields (#3458)
# Motivation

Use the summary fields for Neurons' Fund and direct commitments.

That doesn't mean that the functionality is available in mainnet. The
functionality depends yet on two more fields that are not yet supported.
The fields to give the min and max direct commitment.

# Changes

* Use the summary field for NF commitment in
`getNeuronsFundParticipation`.
* Use the summary field for direct commitment in
`getProjectCommitmentSplit`. We didn't expect it to have a field, that's
why we didn't use a getter initially. But better use it than rely on the
calculation.

# Tests

* Add new fields to test util `createSummary`.
* Tests don't rely on mocking the `getNeuronsFundParticipation` and
instead build the desired summary.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet.