GIX-1860: Refactor PageHeading subtitle slot (#3476)
# Motivation

Avoid the problem with the PageHeading spacing when the subtitle was not
defined.

At the same time, we learned that the PageHeading might get two
subtitles; therefore, a slot without being wrapped makes it more
flexible.

# Changes

* Remove the `h4` wrapping the "subtitle" slot in PageHeading.
* New component "HeadingSubtitle" which has the style of the `h4` inside
the PageHeading.
* Use the new component as subtitle slot for PageHeading in
NnsNeuronPageHeading, SnsNeuronPageHeading and CanisterPageHeading.

# Tests

Only refactor of purely UI components. Functionality didn't change which
is confirmed by tests passing without extra changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.