test: migrate modals to vitest (#3490)
# Motivation

Migrate modal related tests to vitest

# Changes

- migrate test
- add `<svelte:options accessors/>` to `TransactionModal.svelte` to
expose functions
- remove test "should disable next button if network not selected" in
`TransactionModal.spec` **

# TODO

Few modals in canisters, sns and neurons were not migrated.

# Notes

** I don't know why this test was passing before as the dropdown is
automatically selected when an account is entenred. So there is not good
way to test if the "next" button is disabled because the dropdown can be
empty only if the account entered is invalid and if invalid, the "next"
button is disabled anyway.