GetTokens: Don't check SNS balance for ckBTC (#3493)
# Motivation

We have a Get Tokens button which gives either ICP or SNS tokens.
It only gives SNS tokens if an SNS universe is selected and the faucet
address has positive balance for that SNS.
Currently we also try to fetch the SNS faucet balance if the universe is
not an SNS universe, for example with ckBTC.
This results in an error on the console.

# Changes

Only try to fetch SNS faucet balance if the selected universe is an SNS.

# Tests

Manually tried getting ICP while ckBTC or ICP or SNS without faucet
balance universe is select.
And getting SNS tokens while an SNS with balance is selected.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary