test: revert migrated tests (#3500)
# Motivation

Our CI has issue with last few tests that were migrated to vitest. As a
results, lots of build never ends - hangs forever.

Of those problematic tests we identitied two in
https://github.com/dfinity/nns-dapp/actions/runs/6460468191/job/17538215849.

That's why this revert those two PRs for now.

- SnsNeurons.spec.ts

Failed to terminate worker while running
/home/runner/work/nns-dapp/nns-dapp/frontend/src/vitests/lib/pages/SnsNeurons.spec.ts.

- CkBTCTransactionModal.spec.ts

Error: Global crypto was not available and none was provided. Please
inlcude a SubtleCrypto implementation. See
https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto

ReferenceError: localStorage is not defined
 ❯ _openDbStore node_modules/@dfinity/auth-client/src/db.ts:17:29
 ❯ Function.create node_modules/@dfinity/auth-client/src/db.ts:76:4

# Notes

From the two, the first with the "failed to terminate" seems the be the
reason of the ci hanging