GIX-1937: Use min max direct participation fields from aggregator (#3551)
# Motivation

Add functionality after Neurons' Fund enhancements.

In this PR, we use the new fields from the SNS Aggregator when we
convert the data to SnsSummary.

# Changes

* Populate new fields `min_direct_participation_icp_e8s` and
`max_direct_participation_icp_e8s` from the different places into the
SnsSummary. Those fields are in the `Init` from `get_init`, but they are
also in the `init` inside the `swap` and also in the `params` of the
`swap`.

# Tests

* Add the fields in the test case for the NF's enhancements fields. The
other tests have those fields not populated and are converted to `[]` as
it was hardcoded before.

# Todos

- [ ] Add entry to changelog (if necessary).
Not yet necessary.