GIX-1994: Preserve provenance (#3570)
# Motivation
We would like to break updates into smaller chunks. This means that .did
files and derived Rust may not all be from the same IC commit, so we
need to track, for each file, where it came from.

# Changes
* Introduce provenance headers that declare where each .did or derived
rust file came from.


# Tests
* CI tests that generating the current .did and derived rust files
yields the same result as the committed code.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>