Paginated account store (#3341)
# Motivation
We have a few huge accounts and many tiny accounts. The current fixed
size stable memory structures are ill suited to this. Work _is_
progressing on unbounded stable memory data structures, however for the
time being we cannot assume that they will be stable and well tested by
the time we migrate. Until then, we have to split our data into pages
and store them in a "classic" stable memory map.

It is not impossible that we will be able to switch to the unbounded
maps. It is too early to tell.

The plan is to proceed with our own pagination and then, when everything
else is ready, potentially switch to the new stable structures.

# Changes
* Define how accounts will be serialized and stored in pages.
* Create a mock data store with this pagination for unit tests.

# Tests
* There are tests that save and restore data from the mock account
store.

# Todos

- [ ] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>