GIX-1982: Tokens page layout (#3595)
# Motivation

Implement the new Tokens page.

In this PR, setup the layout as expected in Figma.

This is the first time for this layout, but it seems to follow the rules
of the current detail pages. Therefore, I applied the same rules and
created a ticket to confirm them with Artem once he's back.

# Changes

* New component `MainWrapper.svelte`. For now used only in the Tokens
page and therefore only in `components/tokens`. It might be moved in the
future.
* New i18n key for the Tokens page title.
* New `MainWrapper` in the `Tokens` page.
* Use the new i18n key for LayoutList in tokens layout.

# Tests

* Add test for tokens layout.

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary.