GIX-1910: Add max Neurons' Fund participation if present (#3606)
# Motivation

With the new NF enhancements, the maximum participation of the NF is set
at the beginning of the swap. But it's not the same as the current
participation.

Add the limit of the NF participation in the static data.

# Changes

* Add a getter for the maximum NF commitment.
* Use the getter in the ProjectSwapDetails and render the value if
present.
* Add a new i18n key for the new value.
* Change the labels for "Maximum Commitment". I believe that if we add a
new "Maximum", it made sense to specify each of the maximum and minimums
already present.

# Tests

* Add a new field in createSummary helper.
* Add a new test in ProjectSwapDetails.spec for new field.

# Todos

- [x] Add entry to changelog (if necessary).