Delete the nightly II updates (#3610)
# Motivation
Currently the nightly job attempts to update II. This is at odds with
the way other updates are run, and doesn't allow the update job to be
run on its own.

It is not clear to me that we need II in the NNS-dapp at all. Do we need
more than snsdemo provides? Maybe just corresponding .did files? If we
can just delete, that would be great. If not, let's move.

# Changes
- Delete the II update

# Tests
- No tests for no code.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>