Script to deploy nns-dapp on DevEnv (#3617)
# Motivation

It is possible to run a local replica on a
[DevEnv](https://dfinity-lab.gitlab.io/private/k8s/k8s/#/bases/apps/devenv/)
such that it is accessible from other machines on DFINITY VPN.
But for nns-dapp to work, it needs different arguments than when
accessing directly on your local machine.

# Changes

Add a script to deploy nns-dapp to a local replica on a DevEnv with the
correct args to access it from other machines.

# Tests

Tested manually to deploy
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network

# Todos

- [x] Add entry to changelog (if necessary).