Use page objects in CkBTCWallet.spec.ts (#3620)
# Motivation

Making the test more readable.

# Changes

1. Use page objects in `CkBTCWallet.spec.ts`.
2. Add test IDs where necessary.
3. Add necessary methods to exiting page objects.
4. Add new page objects.
5. In
`frontend/src/tests/page-objects/SelectDestinationAddress.page-object.ts`
check whether the dropdown input is present before trying to check the
state of the checkbox. In some contexts there is no checkbox and if the
input is already the type of input we need (dropdown vs. text), this is
fine.
6. Add missing types to
`frontend/src/tests/page-objects/WalletPageHeader.page-object.ts`
7. Add `innerHtmlForDebugging` to PageObjectElement to help probe the
state of an element when debugging.

# Tests

yes

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary