Use transaction type Transfer for Send/Receive derived from TransferFrom (#3634)
# Motivation


[Here](https://github.com/dfinity/nns-dapp/blob/5642420e1cadc0fe4c80c11ed2be9d34293c8630/rs/backend/src/accounts_store.rs#L794)
we treat TransferFrom the same as Transfer and turn into a `Send` or a
`Receive`. But
[here](https://github.com/dfinity/nns-dapp/blob/5642420e1cadc0fe4c80c11ed2be9d34293c8630/rs/backend/src/accounts_store.rs#L821)
we pass the transaction type which might be TransferFrom . But [in the
frontend](https://github.com/dfinity/nns-dapp/blob/5642420e1cadc0fe4c80c11ed2be9d34293c8630/frontend/src/lib/canisters/nns-dapp/nns-dapp.idl.js#L87),
we don't know TransactionType TransferFrom and thus it fails to decode
the transaction.

# Changes

Use transaction type `Transfer` on the `Send` or `Receive` created from
a `TransferFrom`.

# Tests

Added a test with a `TransferFrom` and check that it get transaction
type `Transfer`.

# Todos

- [x] Add entry to changelog (if necessary).