Add convertCkBTCToBtcIcrc2 to ckbtc-convert.services (#3651)
# Motivation

We want to use ICRC-2 approve transfer for BTC withdrawal.
This PR adds a service function `convertCkBTCToBtcIcrc2` to be used in
`CkBTCTransactionModal.svelte` in a later PR.

# Changes

1. Add `convertCkBTCToBtcIcrc2` in
`frontend/src/lib/services/ckbtc-convert.services.ts` which approves the
transfer by the minter canister and than asks the minter to make the
withdrawal.
2. Add a new `APPROVE_TRANSFER` step in `ConvertBtcStep`. This will be
used in a later PR to render the ICRC-2 steps correctly if the feature
flag is enabled.
3. Separate a `reload` function from `retrieveBtcAndReload` to reuse in
`convertCkBTCToBtcIcrc2`.

# Tests

Added unit tests.
Manually tested in a separate branch which has the service function
hooked up already.

# Todos

- [ ] Add entry to changelog (if necessary).
Will add when hooked up.