Create dependabot.yml (#3668)
# Motivation
We would like to keep the Rust dependencies up to date. Some are simple
to manage, third party packages with semantic versions and these can be
taken care of by dependabot. Others, in particular updates of crates in
the IC repo, require a tailored approach. This PR addresses simple third
party packages.

Note: Dependabot opens PRs when updates are available.

### References
- https://dfinity.atlassian.net/browse/GIX-1964
- https://dfinity.atlassian.net/browse/GIX-1967

# Changes
- Enable dependabot for rust crate updates.

# Tests

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>