Update ic workflow (#3660)
# Motivation
We would like to keep the NNS canister declarations up to date.

# Changes
* Add a workflow that checks nightly for new mainnet releases and, if it
finds one, opens a PR to update the nns-dapp declarations.
* Note: IC repo releases are approximately weekly but it varies and the
day of release is also not 100% consistent, so it is worth checking
nightly.
* Note: The workflow can also be run manually to set the declarations to
a particular release or commit.
* Removed the test that the aggregator bindings are in sync with the
declarations. Our latest policy is that parts such as the aggregator
bindings shall be updated independently.

# Tests
* This branch created [this
PR](https://github.com/dfinity/nns-dapp/pull/3664), which passes all
tests apart from the one removed in this PR.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>