update: Use checksummed schema (#3641)
# Motivation
Context: According to the schema migration plan, the schema should be
recorded in the first bytes of stable memory. The schema version is
checksummed, so if there is no valid checksum we assume that we have
stable memory according to the old format that is `<len:u32>||<data
serialized as candid>`.

Currently this is not implemented: The placeholder function that gets
the schema currently always returns None without looking at the stable
memory.

# Changes
- Read the schema version from stable memory.

# Tests
- The downgrade upgrade test will verify that the `None` schema is still
recognized correctly.
- Other serializations are not defined yet.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy <max@winning.black>