Load accounts after withdrawing BTC with ICRC-2 (#3686)
# Motivation

In the "legacy" BTC withdrawal flow (without ICRC-2), one of the steps
is transferring ckBTC to the minter withdrawal account.
The service function to transfer the ckBTC automatically reloads the
account (and thus balance) after doing the transfer.
With the ICRC-2 flow, it's the minter that makes the transfer so we
don't automatically reload accounts.
So we show explicitly reload accounts at the end of the ICRC-2 flow.

# Changes

1. Add a parameter to the `reload` function to indicate if accounts
should be loaded.
2. Pass `true` in the ICRC-2 flow and `false` in the "legacy" flow.

# Tests

In the unit test, mock the call to `loadCkBTCAccounts` and expect it to
be called or not to be called.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary