Fix: Canister title when not controller (#3688)
# Motivation

There was a bug where the user would see a loading skeleton instead of
the canister name when they were not the controllers.

The cause of this bug was a change in the error message from the IC
Management canister when the caller is not the controller.

The error message is parsed and expects a certain structure and wording.

I also changed splitting the error message to get the error code line.
This was not the cause of the bug, because the code checked whether 403
was in the error message. But that would not be accurate because 403
could be part of a request id.

# Changes

* Change the `mapError` function for the IC Management canister.

# Tests

* Added a test with an up to date error message.

# Todos

- [x] Add entry to changelog (if necessary).